Move realizability logic from Types to TypeOps.
Types is already big enough.
